Added monorepo examples
