+++ examples/src/camera/bin/run.rs
