+++ examples/src/camera/src/camera.pax
@@ -1,0 +1,19 @@
+
+
+//Camera
+<Frame width=100% height=100% @click=handle_click >
+    <Group transform={
+        Transform2D::scale((100*zoom)%, (100*zoom)%) *
+        Transform2D::translate((pan_x)px, (pan_y)px)
+    }>
+        <Rectangle width=100px height=100px fill={Fill::Solid(Color::rgb(100.0, 0, 0))} transform={Transform2D::translate(0px,0px)} />
+        <Rectangle width=100px height=100px fill={Fill::Solid(Color::rgb(0, 100.0, 0))} transform={Transform2D::translate(0px,200px)} />
+        <Rectangle width=100px height=100px fill={Fill::Solid(Color::rgb(0, 0, 100.0))} transform={Transform2D::translate(200px,0px)} />
+        <Ellipse width=100px height=100px fill={Color::rgb(0, 50.0, 50.0)} transform={Transform2D::translate(200px,200px)} />
+    </Group>
+</Frame>
+
+
+@handlers {
+    did_mount: handle_did_mount,
+}
