+++ examples/src/color-grid/bin/run.rs
