+++ examples/src/color-grid/src/color-grid.pax
