+++ examples/src/fireworks/bin/run.rs
