+++ examples/src/grids/src/lib.rs
@@ -1,0 +1,12 @@
+use pax_lang::api::*;
+use pax_lang::*;
+use pax_std::primitives::{Ellipse, Frame, Group, Path, Rectangle, Text};
+use pax_std::components::{Stacker};
+use pax_std::types::{Color, StackerDirection};
+
+
+#[derive(Pax)]
+#[main]
+#[file("grids.pax")]
+pub struct Grids {
+}
