+++ examples/src/hello-rgb/bin/run.rs
