+++ examples/src/hello-rgb/src/hello-rgb.pax
@@ -1,0 +1,14 @@
+<Group @wheel=handle_scroll >
+  for i in 0..5 {
+        <Rectangle id=id_test/>
+    }
+</Group>
+
+@settings {
+    #id_test {
+        height: 50%,
+         width: 50%,
+        transform: {Transform2D::translate(50%, 50%) * Transform2D::anchor(50%, 50%) * Transform2D::rotate(Rotation::Percent(rotation))},
+        fill: {Fill::Solid(Color::rgb(0.4,0.5,0))},
+    }
+}
