--- pax-example/components/fireworks.pax
@@ -1,15 +1,0 @@
-<Group @scroll=self.handle_scroll >
-    for i in 0..60 {
-        <Rectangle fill={Fill::Solid(Color::hlc(ticks + i * 360.0 / 30.0, 75.0, 150.0))} width=300px height=300px transform={
-            Transform2D::anchor(50%, 50%)
-            * Transform2D::translate(50%, 50%)
-            * Transform2D::rotate((i + 2) * rotation + ticks / 1000.0)
-            * Transform2D::scale(0.75 + (i * rotation), 0.75 + (i * rotation))
-            * Transform2D::scale(1 - ((rotation / 5) + i / 1000.0), 1 - ((rotation / 5) + i / 1000.0))
-        } corner_radii= { RectangleCornerRadii::radii(50.0,100.0,200.0,300.0)} />
-    }
-</Group>
-
-@handlers {
-    will_render: handle_will_render
-}
