--- pax-example/components/hello_rgb.pax
@@ -1,19 +1,0 @@
-<Ellipse @scroll=self.handle_scroll @click=self.handle_click id=id_test class=class_a class=class_b />
-
-@handlers {
-        click: [self.handle_global_click],
-        scroll: self.handle_global_scroll,
-}
-
-@settings {
-    #id_test {
-        height: 50%,
-         width: 50%,
-    }
-    .class_a {
-        fill: {Color::rgb(0.4,0.5,0)},
-    }
-    .class_b {
-        transform: {Transform2D::translate(50%, 50%) * Transform2D::anchor(50%, 50%) * Transform2D::rotate(rotation)},
-    }
-}
